+++ src/components/AppForm.spec.js
@@ -1,0 +1,5 @@
+describe('GameForm', () => {
+  it.todo('renders a form with an input and a button')
+
+  it.todo('applies the list of names to the submit callback')
+})
