+++ src/components/AppHeader.spec.js
@@ -1,0 +1,5 @@
+import AppHeader from './AppHeader'
+
+describe('AppHeader', () => {
+  it.todo('should render a component from props')
+})
