+++ src/components/AppNavigation.spec.js
@@ -1,0 +1,5 @@
+import AppNavigation from './AppNavigation'
+
+describe('AppNavigation', () => {
+  it.todo('should render a navigation with two links')
+})
