+++ src/components/Button.spec.js
@@ -1,0 +1,5 @@
+import Button from './Button'
+
+describe('Button', () => {
+  it.todo('shows a different color when isActive is true')
+})
